# MAINT: sparse: optimize DIA sum/diagonal, csgraph.laplacian