This PR adds specialized `sum` and `diagonal` methods for DIA sparse matrices. It also defers the conversion to COO format in `csgraph.laplacian` where possible, which yields a pretty nice speedup for DIA inputs without harming other cases.

I've also added benchmarks to document the performance increase. Here are the significant changes:
```
BENCHMARK                                    BEFORE      AFTER     FACTOR
...cian.time_laplacian(30, 'dia', False)   372.06μs   171.08μs   0.45982349x
...ian.time_laplacian(300, 'dia', False)   593.68μs   227.67μs   0.38348440x
...ian.time_laplacian(900, 'dia', False)     1.04ms   330.61μs   0.31713714x
sparse.Sum.time_sum_axis0(0.5, 'dia')      124.03ms    15.37ms   0.12388698x
sparse.Sum.time_sum_axis0(0.1, 'dia')      130.85ms    15.54ms   0.11876036x
sparse.Sum.time_sum_axis0(0.01, 'dia')     124.67ms    13.94ms   0.11180419x
...e.Diagonal.time_diagonal(0.01, 'dia')    21.53ms    17.11μs   0.00079465x
...se.Diagonal.time_diagonal(0.1, 'dia')    30.09ms    17.94μs   0.00059623x
...se.Diagonal.time_diagonal(0.5, 'dia')    50.90ms    18.32μs   0.00035993x
```