@perimosocordiae 2016-03-24
Rebased and slightly tweaked. I simplified the copy logic a bit, and fixed a latent bug related to result dtypes that was exposed by #5907. Should be good to go now.